pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spellbook-vault"
version = "0.1.0"
description = "Personal knowledge vault for Claude Code with subagentic workflows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Samuel Bunce" }
]
keywords = ["claude", "ai", "knowledge-management", "vault"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
sb = "spellbook.cli:cli"

[project.urls]
Homepage = "https://github.com/SBunce92/spellbook"
Documentation = "https://github.com/SBunce92/spellbook/docs"
Repository = "https://github.com/SBunce92/spellbook"

[tool.hatch.build.targets.wheel]
packages = ["src/spellbook"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
